o kadar çok şey yaptım ki anlat anlat bitmez, view models are added and they are connected to perform db load operation of a tracked activity data

+++ app/src/main/java/com/cse234/assets/data/ActivityData.kt
@@ -6,5 +6,5 @@ data class ActivityData(
     val startTime: Long,
     val endTime: Long,
     val duration: Long,
-    val distance: Float
+    val distance: Double
 )
